+++ resources/aws_quicksight_analysis.rb
@@ -16,7 +16,7 @@ property :analysis_id, String,
          required: true,
          callbacks: {
            "analysis_id is not a String" => lambda { |v| v.is_a? String },
-           "analysis_id needs to be 1..2048 characters" => lambda { |v| v.length >= 1 && v.length <= 2048 },
+           "analysis_id needs to be 1..512 characters" => lambda { |v| v.length >= 1 && v.length <= 512 },
            "analysis_id must match pattern [\w\-]+" => lambda { |v| v =~ Regexp.new("/[\w\-]+/") },
          },
          description: ""
@@ -30,30 +30,31 @@ property :aws_account_id, String,
          },
          description: ""
 
-property :errors, Array,
+property :definition, Hash,
          callbacks: {
-           "errors is not a Array" => lambda { |v| v.is_a? Array },
+           "Subproperty `DataSetIdentifierDeclarations` is not a Array" => lambda { |v| v[:DataSetIdentifierDeclarations].is_a? Array },
+           "Subproperty `Sheets` is not a Array" => lambda { |v| v[:Sheets].is_a? Array },
+           "Subproperty `CalculatedFields` is not a Array" => lambda { |v| v[:CalculatedFields].is_a? Array },
+           "Subproperty `ParameterDeclarations` is not a Array" => lambda { |v| v[:ParameterDeclarations].is_a? Array },
+           "Subproperty `FilterGroups` is not a Array" => lambda { |v| v[:FilterGroups].is_a? Array },
+           "Subproperty `ColumnConfigurations` is not a Array" => lambda { |v| v[:ColumnConfigurations].is_a? Array },
          },
-         description: <<~'DESCRIPTION'
-           <p>Errors associated with the analysis.</p>
-         DESCRIPTION
+         description: ""
 
 property :name, String,
          name_property: true,
+         required: true,
          callbacks: {
            "name is not a String" => lambda { |v| v.is_a? String },
            "name needs to be 1..2048 characters" => lambda { |v| v.length >= 1 && v.length <= 2048 },
-           "name must match pattern [\u0020-\u00FF]+" => lambda { |v| v =~ Regexp.new("/[\u0020-\u00FF]+/") },
          },
-         description: <<~'DESCRIPTION'
-           <p>The descriptive name of the analysis.</p>
-         DESCRIPTION
+         description: ""
 
 property :parameters, Hash,
          callbacks: {
            "Subproperty `StringParameters` is not a Array" => lambda { |v| v[:StringParameters].is_a? Array },
-           "Subproperty `DecimalParameters` is not a Array" => lambda { |v| v[:DecimalParameters].is_a? Array },
            "Subproperty `IntegerParameters` is not a Array" => lambda { |v| v[:IntegerParameters].is_a? Array },
+           "Subproperty `DecimalParameters` is not a Array" => lambda { |v| v[:DecimalParameters].is_a? Array },
            "Subproperty `DateTimeParameters` is not a Array" => lambda { |v| v[:DateTimeParameters].is_a? Array },
          },
          description: ""
@@ -62,34 +63,29 @@ property :permissions, Array,
          callbacks: {
            "permissions is not a Array" => lambda { |v| v.is_a? Array },
          },
-         description: <<~'DESCRIPTION'
-           <p>A structure that describes the principals and the resource-level permissions on an
-      analysis. You can use the <code>Permissions</code> structure to grant permissions by
-      providing a list of AWS Identity and Access Management (IAM) action information for each
-      principal listed by Amazon Resource Name (ARN). </p>
-            <p>To specify no permissions, omit <code>Permissions</code>.</p>
-         DESCRIPTION
+         description: ""
 
 property :source_entity, Hash,
-         required: true,
+         description: ""
+
+property :status, Hash,
+         callbacks: {
+           "status is not a String" => lambda { |v| v.is_a? String },
+           "statusis not one of `CREATION_IN_PROGRESS`, `CREATION_SUCCESSFUL`, `CREATION_FAILED`, `UPDATE_IN_PROGRESS`, `UPDATE_SUCCESSFUL`, `UPDATE_FAILED`, `DELETED`" => lambda { |v| %w{CREATION_IN_PROGRESS CREATION_SUCCESSFUL CREATION_FAILED UPDATE_IN_PROGRESS UPDATE_SUCCESSFUL UPDATE_FAILED DELETED}.include? v },
+         },
          description: ""
 
 property :tags, Array,
          callbacks: {
            "tags is not a Array" => lambda { |v| v.is_a? Array },
          },
-         description: <<~'DESCRIPTION'
-           <p>Contains a map of the key-value pairs for the resource tag or tags assigned to the
-      analysis.</p>
-         DESCRIPTION
+         description: ""
 
 property :theme_arn, String,
          callbacks: {
            "theme_arn is not a String" => lambda { |v| v.is_a? String },
          },
-         description: <<~'DESCRIPTION'
-           <p>The ARN of the theme of the analysis.</p>
-         DESCRIPTION
+         description: ""
 
 # API URLs and mappings
 rest_api_collection "/AWS::QuickSight::Analysis"
@@ -98,11 +94,12 @@ rest_api_document "/AWS::QuickSight::Analysis"
 rest_property_map({
   analysis_id:    "AnalysisId",
   aws_account_id: "AwsAccountId",
-  errors:         "Errors",
+  definition:     "Definition",
   name:           "Name",
   parameters:     "Parameters",
   permissions:    "Permissions",
   source_entity:  "SourceEntity",
+  status:         "Status",
   tags:           "Tags",
   theme_arn:      "ThemeArn",
 })